Reorder methods to improve readability
